user test is done untill delete

+++ pages/api/user/[userID]/index.api.ts
@@ -2,11 +2,17 @@ import { NextApiRequest, NextApiResponse } from "next";
 import { z, ZodError } from "zod";
 import { prisma } from "../../../../prisma/db";
 
+class IntegrityError extends Error {
+	constructor() {
+		super();
+	}
+}
+
 const putUser = z.object({
 	image: z.string().optional(),
 	name: z.string().optional(),
 });
-export default async function handler(
+export default async function userIdHandler(
 	req: NextApiRequest,
 	res: NextApiResponse
 ) {
@@ -31,9 +37,17 @@ export default async function handler(
 			res.status(200).json(user);
 		}
 		if (req.method === "DELETE") {
+			const getBooksUserBorrowed = await prisma.book.findFirst({
+				where: {
+					borrowerId: userId,
+				},
+			});
+			if (getBooksUserBorrowed) {
+				throw new IntegrityError();
+			}
 			const deleteUser = await prisma.user.delete({
 				where: {
-					identifier: userId,
+					identifier: "003b79b4-e232-41ef-a9a4-72dbc1d3cea5",
 				},
 			});
 			res.status(200).json(deleteUser);
@@ -44,10 +58,17 @@ export default async function handler(
 				message: "User doesnt exist",
 				error: err,
 			});
+		} else if (err instanceof IntegrityError) {
+			res.status(422).send({
+				message:
+					"You have to return the borrowed books before deleting your account",
+				error: err,
+			});
+		} else {
+			res.status(404).send({
+				message: "Looks like something went wrong. Please try again.",
+				error: err,
+			});
 		}
-		res.status(404).send({
-			message: "Looks like something went wrong. Please try again.",
-			error: err,
-		});
 	}
 }
